--- src/util.js
@@ -1,35 +1,0 @@
-const custTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
-
-export const formatTime = (date) => {
-  const hours = custTimeFormat(date.getHours() % 12);
-  const minutes = custTimeFormat(date.getMinutes());
-  const interval = date.getHours() < 11 ? `AM` : `PM`;
-
-  return `${hours}: ${minutes} ${interval}`;
-};
-
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
