+++ src/utils/render.js
@@ -1,0 +1,40 @@
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export const createElement = (template) => {
+  const parent = document.createElement(`div`);
+  parent.innerHTML = template;
+
+  return parent.firstChild;
+};
+
+export const render = (container, element, place) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+
+    case RenderPosition.BEFOREEND:
+      container.append(element);
+      break;
+  }
+};
+
+export const remove = (component) => {
+  component.getElement().remove();
+  component.remove();
+};
+
+export const replace = (newComponent, oldComponent) => {
+  const parentElement = oldComponent.getElement().parentElement;
+  const newElement = newComponent.getElement();
+  const oldElement = oldComponent.getElement();
+
+  const isElementExists = !!(parentElement && newElement && oldElement);
+
+  if (isElementExists) {
+    parentElement.replaceChild(newElement, oldElement);
+  }
+};
